% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalan_date.R
\name{catalan_date}
\alias{catalan_date}
\title{Make Catalan date}
\usage{
catalan_date(month = TRUE, new_line = TRUE)
}
\arguments{
\item{month}{Whether it should be at the month level}

\item{new_line}{Whether a line break should be inserted}
}
\value{
Function
}
\description{
Create a Catalan date
}
