% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicates_database.R
\name{delete_duplicates_database}
\alias{delete_duplicates_database}
\title{Delete duplicates}
\usage{
delete_duplicates_database()
}
\value{
A set up database
}
\description{
Delete duplicates in the database
}
