% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_data.R
\name{get_google_data}
\alias{get_google_data}
\title{Get google data}
\usage{
get_google_data(
  name = c("parlament", "politicians", "newspapers", "socialists"),
  url = NULL
)
}
\arguments{
\item{name}{The name of the sheet (one of "diputats")}

\item{url}{The url of the sheet}
}
\value{
a tibble
}
\description{
Get a table from google sheets
}
