% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeline_by_date.R
\name{get_timeline_by_date}
\alias{get_timeline_by_date}
\title{Get timeline by date}
\usage{
get_timeline_by_date(user, date, n = 3200)
}
\value{
A dataframe
}
\description{
Get a timeline of a twitter user(s), but go back as far as a certain date (as opposed to the typical use case of simply getting the most recent n tweets). Note that the function in many cases retrieves data OLDER than the date argument.
}
