% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_data_from_database.R
\name{get_twitter_data_from_database}
\alias{get_twitter_data_from_database}
\title{Get twitter data from database}
\usage{
get_twitter_data_from_database(people = "joethebrew")
}
\arguments{
\item{people}{user handles}
}
\value{
a tibble
}
\description{
Get all tweets for a subset of people from the local twitter database
}
