% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Round to nearest multiple of base}
\usage{
mround(x, base)
}
\arguments{
\item{x}{A numeric vector}

\item{base}{Round to the nearest multiple of this}
}
\value{
A numeric vector
}
\description{
Round percentages to get a whole integer
}
