% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_percent.R
\name{round_percent}
\alias{round_percent}
\title{Round percent}
\usage{
round_percent(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Round percentages to get a whole integer
}
