% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Scrape data using twint}
\usage{
scrape(search_string, file_path = "twitter_data.csv",
  start_date = NULL, end_date = NULL)
}
\arguments{
\item{search_string}{A Character string}

\item{file_path}{Path to file}

\item{start_date}{The date to start with}

\item{end_date}{The date to end with}
}
\value{
A csv file
}
\description{
Use twint to get data
}
