% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vilaweb.R
\name{theme_vilaweb}
\alias{theme_vilaweb}
\title{vilaweb theme}
\usage{
theme_vilaweb(base_size = 15, base_family = "Avenir", y_comma = TRUE,
  subtitle_family = "Avenir", axis_family = "Avenir")
}
\arguments{
\item{base_size}{The size of the base font}

\item{base_family}{The font used in the legend and title}

\item{y_comma}{Whether to add comma seperators every 3 digits to a continuous
y axis}

\item{subtitle_family}{The font used in the subtitle}

\item{axis_family}{The font used in the axes}
}
\value{
A list of options compatible with \code{ggplot()}
}
\description{
Apply the vilaweb look to a ggplot2-generated visualization
}
