% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update the local twitter database}
\usage{
update_database(people = NULL, get_new = FALSE,
  delete_duplicates = TRUE)
}
\arguments{
\item{people}{A character vector of twitter accounts. If null, will do existing people}

\item{get_new}{Get new data. If TRUE, will fetch all tweets again.}

\item{delete_duplicates}{If TRUE, will export the data to R, remove duplicates, then overwrite the db}
}
\value{
A set up database
}
\description{
Update the local twitter database, which should have been set up via inst/rmd/set_up_database/set_up_database.R
}
