% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update the local twitter database}
\usage{
update_database(people = NULL, get_new = FALSE,
  delete_duplicates = TRUE, only_new_people = FALSE,
  after = "2000-01-01", until = as.character(Sys.Date() + 1),
  force_old = FALSE)
}
\arguments{
\item{people}{A character vector of twitter accounts. If null, will do existing people}

\item{get_new}{Get new data. If TRUE, will fetch all tweets again.}

\item{delete_duplicates}{If TRUE, will export the data to R, remove duplicates, then overwrite the db}

\item{only_new_people}{Whether to only fetch people who are NOT yet in the database (ie, automatically filters the people argument)}

\item{after}{The date after which tweets should be retrieved. Default to 2000-01-01}

\item{until}{The date until which tweets will be retrieved}

\item{force_old}{Set to TRUE if retrieving tweets older than what is already in the database (only for exceptional cases)}
}
\value{
A set up database
}
\description{
Update the local twitter database, which should have been set up via inst/rmd/set_up_database/set_up_database.R
}
